Change jooqcodegen path
