+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/DefaultCatalog.java
@@ -2,7 +2,7 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq;
+package edu.java.domain.jooqcodegen;
 
 import java.util.Arrays;
 import java.util.List;
