+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/DefaultSchema.java
@@ -2,13 +2,13 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq;
+package edu.java.domain.jooqcodegen;
 
-import edu.java.domain.jooq.tables.GithubLinks;
-import edu.java.domain.jooq.tables.Links;
-import edu.java.domain.jooq.tables.StackoverflowLinks;
-import edu.java.domain.jooq.tables.UserTrackedLinks;
-import edu.java.domain.jooq.tables.Users;
+import edu.java.domain.jooqcodegen.tables.GithubLinks;
+import edu.java.domain.jooqcodegen.tables.Links;
+import edu.java.domain.jooqcodegen.tables.StackoverflowLinks;
+import edu.java.domain.jooqcodegen.tables.UserTrackedLinks;
+import edu.java.domain.jooqcodegen.tables.Users;
 import java.util.Arrays;
 import java.util.List;
 import javax.annotation.processing.Generated;
