+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/Keys.java
@@ -2,18 +2,18 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq;
+package edu.java.domain.jooqcodegen;
 
-import edu.java.domain.jooq.tables.GithubLinks;
-import edu.java.domain.jooq.tables.Links;
-import edu.java.domain.jooq.tables.StackoverflowLinks;
-import edu.java.domain.jooq.tables.UserTrackedLinks;
-import edu.java.domain.jooq.tables.Users;
-import edu.java.domain.jooq.tables.records.GithubLinksRecord;
-import edu.java.domain.jooq.tables.records.LinksRecord;
-import edu.java.domain.jooq.tables.records.StackoverflowLinksRecord;
-import edu.java.domain.jooq.tables.records.UserTrackedLinksRecord;
-import edu.java.domain.jooq.tables.records.UsersRecord;
+import edu.java.domain.jooqcodegen.tables.GithubLinks;
+import edu.java.domain.jooqcodegen.tables.Links;
+import edu.java.domain.jooqcodegen.tables.StackoverflowLinks;
+import edu.java.domain.jooqcodegen.tables.UserTrackedLinks;
+import edu.java.domain.jooqcodegen.tables.Users;
+import edu.java.domain.jooqcodegen.tables.records.GithubLinksRecord;
+import edu.java.domain.jooqcodegen.tables.records.LinksRecord;
+import edu.java.domain.jooqcodegen.tables.records.StackoverflowLinksRecord;
+import edu.java.domain.jooqcodegen.tables.records.UserTrackedLinksRecord;
+import edu.java.domain.jooqcodegen.tables.records.UsersRecord;
 import javax.annotation.processing.Generated;
 import org.jooq.ForeignKey;
 import org.jooq.TableField;
