+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/enums/LinkTypeEnum.java
@@ -2,9 +2,9 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.enums;
+package edu.java.domain.jooqcodegen.enums;
 
-import edu.java.domain.jooq.DefaultSchema;
+import edu.java.domain.jooqcodegen.DefaultSchema;
 import javax.annotation.processing.Generated;
 import org.jooq.Catalog;
 import org.jooq.EnumType;
@@ -21,17 +21,15 @@ import org.jooq.Schema;
     comments = "This class is generated by jOOQ"
 )
 @SuppressWarnings({"all", "unchecked", "rawtypes", "this-escape"})
-public enum UserStatusEnum implements EnumType {
+public enum LinkTypeEnum implements EnumType {
 
-    base("base"),
+    github("github"),
 
-    track_link("track_link"),
-
-    untrack_link("untrack_link");
+    stackoverflow("stackoverflow");
 
     private final String literal;
 
-    private UserStatusEnum(String literal) {
+    private LinkTypeEnum(String literal) {
         this.literal = literal;
     }
 
@@ -47,7 +45,7 @@ public enum UserStatusEnum implements EnumType {
 
     @Override
     public String getName() {
-        return "USER_STATUS_ENUM";
+        return "LINK_TYPE_ENUM";
     }
 
     @Override
@@ -58,7 +56,7 @@ public enum UserStatusEnum implements EnumType {
     /**
      * Lookup a value of this EnumType by its literal
      */
-    public static UserStatusEnum lookupLiteral(String literal) {
-        return EnumType.lookupLiteral(UserStatusEnum.class, literal);
+    public static LinkTypeEnum lookupLiteral(String literal) {
+        return EnumType.lookupLiteral(LinkTypeEnum.class, literal);
     }
 }
