+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/enums/UserStatusEnum.java
@@ -2,9 +2,9 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.enums;
+package edu.java.domain.jooqcodegen.enums;
 
-import edu.java.domain.jooq.DefaultSchema;
+import edu.java.domain.jooqcodegen.DefaultSchema;
 import javax.annotation.processing.Generated;
 import org.jooq.Catalog;
 import org.jooq.EnumType;
@@ -21,15 +21,17 @@ import org.jooq.Schema;
     comments = "This class is generated by jOOQ"
 )
 @SuppressWarnings({"all", "unchecked", "rawtypes", "this-escape"})
-public enum LinkTypeEnum implements EnumType {
+public enum UserStatusEnum implements EnumType {
 
-    github("github"),
+    base("base"),
 
-    stackoverflow("stackoverflow");
+    track_link("track_link"),
+
+    untrack_link("untrack_link");
 
     private final String literal;
 
-    private LinkTypeEnum(String literal) {
+    private UserStatusEnum(String literal) {
         this.literal = literal;
     }
 
@@ -45,7 +47,7 @@ public enum LinkTypeEnum implements EnumType {
 
     @Override
     public String getName() {
-        return "LINK_TYPE_ENUM";
+        return "USER_STATUS_ENUM";
     }
 
     @Override
@@ -56,7 +58,7 @@ public enum LinkTypeEnum implements EnumType {
     /**
      * Lookup a value of this EnumType by its literal
      */
-    public static LinkTypeEnum lookupLiteral(String literal) {
-        return EnumType.lookupLiteral(LinkTypeEnum.class, literal);
+    public static UserStatusEnum lookupLiteral(String literal) {
+        return EnumType.lookupLiteral(UserStatusEnum.class, literal);
     }
 }
