+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/tables/Links.java
@@ -2,17 +2,12 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.tables;
+package edu.java.domain.jooqcodegen.tables;
 
-import edu.java.domain.jooq.DefaultSchema;
-import edu.java.domain.jooq.Keys;
-import edu.java.domain.jooq.enums.LinkTypeEnum;
-import edu.java.domain.jooq.tables.records.LinksRecord;
-import java.time.OffsetDateTime;
-import java.util.Arrays;
-import java.util.List;
-import java.util.function.Function;
-import javax.annotation.processing.Generated;
+import edu.java.domain.jooqcodegen.DefaultSchema;
+import edu.java.domain.jooqcodegen.Keys;
+import edu.java.domain.jooqcodegen.enums.LinkTypeEnum;
+import edu.java.domain.jooqcodegen.tables.records.LinksRecord;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jooq.Field;
@@ -32,6 +27,12 @@ import org.jooq.UniqueKey;
 import org.jooq.impl.DSL;
 import org.jooq.impl.SQLDataType;
 import org.jooq.impl.TableImpl;
+
+import javax.annotation.processing.Generated;
+import java.time.OffsetDateTime;
+import java.util.Arrays;
+import java.util.List;
+import java.util.function.Function;
 
 /**
  * This class is generated by jOOQ.
@@ -72,7 +73,7 @@ public class Links extends TableImpl<LinksRecord> {
      * The column <code>LINKS.LINK_TYPE</code>.
      */
     public final TableField<LinksRecord, LinkTypeEnum> LINK_TYPE = createField(DSL.name("LINK_TYPE"),
-        SQLDataType.VARCHAR.nullable(false).asEnumDataType(edu.java.domain.jooq.enums.LinkTypeEnum.class),
+            SQLDataType.VARCHAR.nullable(false).asEnumDataType(edu.java.domain.jooqcodegen.enums.LinkTypeEnum.class),
         this,
         ""
     );
