+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/tables/StackoverflowLinks.java
@@ -2,11 +2,11 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.tables;
+package edu.java.domain.jooqcodegen.tables;
 
-import edu.java.domain.jooq.DefaultSchema;
-import edu.java.domain.jooq.Keys;
-import edu.java.domain.jooq.tables.records.StackoverflowLinksRecord;
+import edu.java.domain.jooqcodegen.DefaultSchema;
+import edu.java.domain.jooqcodegen.Keys;
+import edu.java.domain.jooqcodegen.tables.records.StackoverflowLinksRecord;
 import java.time.OffsetDateTime;
 import java.util.Arrays;
 import java.util.List;
