+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/tables/UserTrackedLinks.java
@@ -2,11 +2,11 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.tables;
+package edu.java.domain.jooqcodegen.tables;
 
-import edu.java.domain.jooq.DefaultSchema;
-import edu.java.domain.jooq.Keys;
-import edu.java.domain.jooq.tables.records.UserTrackedLinksRecord;
+import edu.java.domain.jooqcodegen.DefaultSchema;
+import edu.java.domain.jooqcodegen.Keys;
+import edu.java.domain.jooqcodegen.tables.records.UserTrackedLinksRecord;
 import java.util.Arrays;
 import java.util.List;
 import java.util.function.Function;
