+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/tables/Users.java
@@ -2,12 +2,12 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.tables;
+package edu.java.domain.jooqcodegen.tables;
 
-import edu.java.domain.jooq.DefaultSchema;
-import edu.java.domain.jooq.Keys;
-import edu.java.domain.jooq.enums.UserStatusEnum;
-import edu.java.domain.jooq.tables.records.UsersRecord;
+import edu.java.domain.jooqcodegen.DefaultSchema;
+import edu.java.domain.jooqcodegen.Keys;
+import edu.java.domain.jooqcodegen.enums.UserStatusEnum;
+import edu.java.domain.jooqcodegen.tables.records.UsersRecord;
 import java.util.Arrays;
 import java.util.List;
 import java.util.function.Function;
@@ -71,7 +71,7 @@ public class Users extends TableImpl<UsersRecord> {
      */
     public final TableField<UsersRecord, UserStatusEnum> USER_STATUS = createField(DSL.name("USER_STATUS"),
         SQLDataType.VARCHAR.nullable(false).defaultValue(DSL.field(DSL.raw("'base'"), SQLDataType.VARCHAR))
-            .asEnumDataType(edu.java.domain.jooq.enums.UserStatusEnum.class),
+            .asEnumDataType(edu.java.domain.jooqcodegen.enums.UserStatusEnum.class),
         this,
         ""
     );
