+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/tables/pojos/Links.java
@@ -2,9 +2,9 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.tables.pojos;
+package edu.java.domain.jooqcodegen.tables.pojos;
 
-import edu.java.domain.jooq.enums.LinkTypeEnum;
+import edu.java.domain.jooqcodegen.enums.LinkTypeEnum;
 import jakarta.validation.constraints.Size;
 import java.beans.ConstructorProperties;
 import java.io.Serializable;
