+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/tables/pojos/UserTrackedLinks.java
@@ -2,7 +2,7 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.tables.pojos;
+package edu.java.domain.jooqcodegen.tables.pojos;
 
 import java.beans.ConstructorProperties;
 import java.io.Serializable;
