+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/tables/pojos/Users.java
@@ -2,9 +2,9 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.tables.pojos;
+package edu.java.domain.jooqcodegen.tables.pojos;
 
-import edu.java.domain.jooq.enums.UserStatusEnum;
+import edu.java.domain.jooqcodegen.enums.UserStatusEnum;
 import java.beans.ConstructorProperties;
 import java.io.Serializable;
 import javax.annotation.processing.Generated;
