+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/tables/records/GithubLinksRecord.java
@@ -2,9 +2,9 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.tables.records;
+package edu.java.domain.jooqcodegen.tables.records;
 
-import edu.java.domain.jooq.tables.GithubLinks;
+import edu.java.domain.jooqcodegen.tables.GithubLinks;
 import java.beans.ConstructorProperties;
 import java.time.OffsetDateTime;
 import javax.annotation.processing.Generated;
@@ -267,7 +267,7 @@ public class GithubLinksRecord extends UpdatableRecordImpl<GithubLinksRecord>
     /**
      * Create a detached, initialised GithubLinksRecord
      */
-    public GithubLinksRecord(edu.java.domain.jooq.tables.pojos.GithubLinks value) {
+    public GithubLinksRecord(edu.java.domain.jooqcodegen.tables.pojos.GithubLinks value) {
         super(GithubLinks.GITHUB_LINKS);
 
         if (value != null) {
