+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/tables/records/LinksRecord.java
@@ -2,10 +2,10 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.tables.records;
+package edu.java.domain.jooqcodegen.tables.records;
 
-import edu.java.domain.jooq.enums.LinkTypeEnum;
-import edu.java.domain.jooq.tables.Links;
+import edu.java.domain.jooqcodegen.enums.LinkTypeEnum;
+import edu.java.domain.jooqcodegen.tables.Links;
 import jakarta.validation.constraints.Size;
 import java.beans.ConstructorProperties;
 import java.time.OffsetDateTime;
@@ -271,7 +271,7 @@ public class LinksRecord extends UpdatableRecordImpl<LinksRecord>
     /**
      * Create a detached, initialised LinksRecord
      */
-    public LinksRecord(edu.java.domain.jooq.tables.pojos.Links value) {
+    public LinksRecord(edu.java.domain.jooqcodegen.tables.pojos.Links value) {
         super(Links.LINKS);
 
         if (value != null) {
