+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/tables/records/StackoverflowLinksRecord.java
@@ -2,9 +2,9 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.tables.records;
+package edu.java.domain.jooqcodegen.tables.records;
 
-import edu.java.domain.jooq.tables.StackoverflowLinks;
+import edu.java.domain.jooqcodegen.tables.StackoverflowLinks;
 import java.beans.ConstructorProperties;
 import java.time.OffsetDateTime;
 import javax.annotation.processing.Generated;
@@ -223,7 +223,7 @@ public class StackoverflowLinksRecord extends UpdatableRecordImpl<StackoverflowL
     /**
      * Create a detached, initialised StackoverflowLinksRecord
      */
-    public StackoverflowLinksRecord(edu.java.domain.jooq.tables.pojos.StackoverflowLinks value) {
+    public StackoverflowLinksRecord(edu.java.domain.jooqcodegen.tables.pojos.StackoverflowLinks value) {
         super(StackoverflowLinks.STACKOVERFLOW_LINKS);
 
         if (value != null) {
