+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/tables/records/UserTrackedLinksRecord.java
@@ -2,9 +2,9 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.tables.records;
+package edu.java.domain.jooqcodegen.tables.records;
 
-import edu.java.domain.jooq.tables.UserTrackedLinks;
+import edu.java.domain.jooqcodegen.tables.UserTrackedLinks;
 import java.beans.ConstructorProperties;
 import javax.annotation.processing.Generated;
 import org.jetbrains.annotations.NotNull;
@@ -170,7 +170,7 @@ public class UserTrackedLinksRecord extends UpdatableRecordImpl<UserTrackedLinks
     /**
      * Create a detached, initialised UserTrackedLinksRecord
      */
-    public UserTrackedLinksRecord(edu.java.domain.jooq.tables.pojos.UserTrackedLinks value) {
+    public UserTrackedLinksRecord(edu.java.domain.jooqcodegen.tables.pojos.UserTrackedLinks value) {
         super(UserTrackedLinks.USER_TRACKED_LINKS);
 
         if (value != null) {
