+++ scrapper/src/main/java/edu/java/domain/jooqcodegen/tables/records/UsersRecord.java
@@ -2,10 +2,10 @@
  * This file is generated by jOOQ.
  */
 
-package edu.java.domain.jooq.tables.records;
+package edu.java.domain.jooqcodegen.tables.records;
 
-import edu.java.domain.jooq.enums.UserStatusEnum;
-import edu.java.domain.jooq.tables.Users;
+import edu.java.domain.jooqcodegen.enums.UserStatusEnum;
+import edu.java.domain.jooqcodegen.tables.Users;
 import java.beans.ConstructorProperties;
 import javax.annotation.processing.Generated;
 import org.jetbrains.annotations.NotNull;
@@ -161,7 +161,7 @@ public class UsersRecord extends TableRecordImpl<UsersRecord> implements Record2
     /**
      * Create a detached, initialised UsersRecord
      */
-    public UsersRecord(edu.java.domain.jooq.tables.pojos.Users value) {
+    public UsersRecord(edu.java.domain.jooqcodegen.tables.pojos.Users value) {
         super(Users.USERS);
 
         if (value != null) {
